fix production websocket tests so they don't depend on the stock market

+++ api/finhub/finhub.go
@@ -35,7 +35,7 @@ type WebsocketResponse struct {
 type Trade struct {
 	Symbol    string  `json:"s"`
 	Price     float32 `json:"p"`
-	Volume    int     `json:"v"`
+	Volume    float32 `json:"v"`
 	Timestamp int64   `json:"t"`
 }
